requirements.txt
django==4.0.4
tensorflow==2.9.0
Pillow==9.1.1
